Assignment 1 - added publisher entity class, repository, and demo data in bootsrapData class as well.

+++ src/main/java/com/example/demo/repository/AuthorRepository.java
@@ -3,6 +3,6 @@ package com.example.demo.repository;
 import com.example.demo.domain.Author;
 import org.springframework.data.repository.CrudRepository;
 
-public interface AuthorRepository extends CrudRepository<Author, Long> {
+public interface AuthorRepository extends CrudRepository<Author, Integer> {
 
 }
